Fix LUX error in IE 11
Fixes #2804 by adding an extra check to make sure that a variable exists before
seeing what its length is.

Also removes an unnecessary use of `JSON.parse` and  `JSON.stringify` which was
there to see what an object contained at a specific point in time - this isn't
required for the production code.