Add validation and tests for emergency banner
Emergency banner can only be one of three types (classes),
therefore a helper has been implemented so that if no class
is provided or if the class is not in the list of types then
an exception will be thrown.

There were no tests for the emergency banner component itself
in static (where the component originally resided), so new
tests have been written.