Refactor styles to meet component standards
Remove the parent selectors in emergency_banner.scss as it
breaks the component principle of isolation. Rewrite classes
to use '--homepage' modifiers and add new 'homepage' param to
the partial which will determine if homepage styles should be
used or not. Add tests for correct classes being appended if
homepage param set on partial.

Remove govuk-link from link in emergency banner partial as
the visited colour was not being applied correctly on some
applications.