Update visual regression testing Node version
The Node version specified for use in the visual regression testing workflow has
been updated to use `lts/*` - this matches what is specified in the continuous
integration workflow in this repo.

The latest version of the Percy CLI requires Node v14 or higher - but the VRT
workflow was locked to Node v12.

At the time of writing, the latest long term support version of Node is 16 - so
this will meet the Percy CLI minimum requirements and allow the visual
regression tests to run.