Check if primary_publisher nil OR empty in meta_tags
Tests were intermittently failing on collections and after
some digging, it was discovered that this was because that
the random schema generated for tests can occasionally
result in primary_publishing_organisation being set to an
empty array. This would cause accessing the key of the first
hash of the array in meta_tags to fail which would mean tests
would fail. This commit adds ".blank?" to the primary_publisher
which means that if the array is empty then an error will not
be caused by this line. This fixes the collections tests but
in theory should also fix other intermittently failing tests
in other applications.