Add extra options for show/hide all sections accordion click
The function that fires an event when the Show/Hide All Sections
button is clicked in an accordion can now be passed optional
tracking options. This means custom dimensions can now be specified
when a user clicks on Show/Hide All Sections. This functionality
has been documented for future reference.

Add new tests for the newly added analytics on the accordion.