Change background colour of local emergency
When adding the emergency banner to the gem, the automatic
accessibility tests pointed out that the contrast between
the background colour and text was not high enough for the
local emergency banner styled emergency banner. After
consulting a designer, a new background colour was determined
which ensures the contrast ratio is sufficiently high.