Refactor external/download link tracking to reduce false positives
The previous implementation was not very strict, for example it was looking for 'http' at the start of a href to determine if it was external or not, however hrefs can omit http/https entirely. Therefore this adds stricter checking for link variations: http, https, www., no www., no http/https, no http/https/www., and schema relative '//' links.