Add white arrow to action link options
For the new cost of living banner, the design used the already existing
action link component but with a white arrow instead of the default
colours that were available (so that it is the same colour as the link
text). The white arrow .svg/.png have been added, the documentation has
been updated with the new option and a test for the new arrow colour
has been added.