Unit test the TableHelper
This adds in unit tests for the helper method of the TableHelper class.

There were no tests before so this just adds tests for the logic defined
by the one public method: .helper.

I did consider adding tests that also exercised the TableBuilder but
this is somewhat tricky as the code expects to receive blocks that make
use of puts (via the erb equivalent of `<%=`). This means they'll need
to be a rather contrived approach to test the whole thing.

Instead I decided it was better to leave this as something tested by the
component as a whole. We can also add unit testing later to the
TableBuilder if it increases in scope.

Co-authored-by: Kevin Dew <kevindew@me.com>