Update jasmine tests
The jasmine tests needed updating to reflect the small amount of new HTML I added to the template, and a few small content changes.

During testing however, I found many anomalies causing the tests to fail. It turns out that the tests were very out of date because they did not reflect the most recent updates to the component currently on the site. This might highlight a potential gap in our development processes.

The fix was to completely replace the old fixture with a new one that fully represents the correct rendered HTML.