WIP
Currently, the description isn't read out by a screen reader when the
radio buttons are focused upon. This commit adds the ability to add a
description to the legend containing the heading.

This is needed as we currently have a question that uses the
description as part of the question. E.g:

Question:
Are you getting any of the following?

Description:
Universal Credit
income-based Jobseeker's Allowance (JSA)
income-related Employment and Support Allowance (ESA)
Pension Credit
Tax credits
Income Support
Housing Benefit