WIP - Fix super nav bug on Chrome/FF
We're rolling back the previous code change which fixed a related bug in Safari but which was now causing a layout issue in Chrome/FF. Instead use matchMedia to check the viewport size in modern browsers and fall back to the previous implementation for legacy browsers as we're not aware of an issue with this implementation there.

Needs more testing done and code comments.