Add links and sections count tracking for Cost of living hub
We need to add tracing to populate custom dimensions 26 (number of sections)
and 27 (number of links) on the Cost of living hub pages as requested by a
performance analyst.

data-track-count="accordionSection" is already added to the accordion sections
https://github.com/alphagov/collections/blob/241b33a/app/views/cost_of_living_landing_page/show.html.erb#L22

data-track-count="contentLink" will be added to links within accordions and
links within inset text on the page.