Changes pretty print to use a rails package rather than JSON converter
We found that the pretty_data function was change `:or` keys to "or" strings. This caused copy and pasting of the code to fail as "or" would cause the component to throw an error.

This change adds the awesome_print package to pretty print the rails code. This is a more stable solution and allows us to use things like `:or` keys.