Add cost of living hub links
New links to the Cost of living hub have been added to the navbar and to
the footer. Both of these elements were designed with a specific number
of children in mind so therefore minor stylistic adjustments had to be
made.

The number of rows in the navbar topics toggle menu has had to be
increased to avoid the splitting of the grid into three columns.

A new selector has been added to the layout footer which ensures that
the number of items in each column is correctly balanced when there is
an odd number of links.