Remove hardcoding of internal domains, download paths, and download domains
This uses a config object to allow anyone to specify a list of internal domains, internal download paths, and dedicated download domains. This also fixes a bug where running the code outside of 'www.gov.uk', such as on integration, would incorrectly classify domains on the same link as external. This was fixed by adding the current hostname as an internal domain.