Use section for emergency banner
When the emergency banner was deployed on GOV.UK, some changes were made
based on the feedback from an accessibility specialist. These changes
involved removing the role attribute from the parent div and changing
the element of the parent div to a section. This is to avoid any
accidental duplication of roles on any page of GOV.UK. In accordance
with this change, the aria-label has been changed to aria-labelledby
instead. Have also added a test to make sure the label is being applied
correctly.

Co-authored-by: Max Froumentin <maxf@users.noreply.github.com>