Reformat component to match new design
The original component used float based layout techniques to achieve a variety of different layouts. This was fairly delicate but worked in the context of the original markup.

With the design now requiring tighter positioning of elements, the legacy setup was not robust enough and needed updating. In particular, the "Is this page useful?" question needed to flow better for foreign languages, where the text often wraps to multiple lines on mobile.

Now the layout uses flexbox to arrange elements more robustly, and improves responsiveness at various screen sizes.