Add option for blue bar on full width public layout
For Browse pages in collections, we need to use a full width layout for
the header to fit the width of the entire page. The design of the page
also still includes the blue bar underneath the navbar. Previously, this
was disabled in full width layout with no option to override. This meant
that in collections, a bespoke implementation of the blue bar was
implemented which had styling discrepencies to the blue bar that appears
as part of public layout. This commit adds an option to have a blue bar
in the full width layout so we can remove the custom blue bar on collections.