Add new link_path_parts object to allow longer strings to be pushed
GA4 limits event parameters to 100 characters, which is an issue for URLs as a lot of them exceed 100 characters. Therefore, this commit will always push an object with the URL path in, split every 100 characters. This will allow us to push 500 characters of path/query params to the dataLayer. It will be reconstructed to a full URL using the link_domain value.