Fix axe-core false positive color contrast test
The "Example" grey box shown for each component preview was causing the axe-core colour-contrast check to report as "incomplete". We would then advise that the component needs to be manually checked.

The "Example" element was being nested within the component itself, so was also being included as part of the accessibility test. The accessibility test would then mark the test as incomplete because the "Example" element uses the :before psuedo class to set both the color and background-color which cannot be tested.