Fix issue with blue action link arrow svg
The tail of the blue arrow svg used for action links was misaligned and
not in the centre of the caret. It was previously using a different SVG
to the other arrows for action link. This was because unlike the other
SVGs the arrow for the blue svg is transparent. To fix the styling issue
the blue arrow now uses the svg that the other arrows use but has a mask
so it remains transparent as it was before.,