Fix performance issue with data-ga4-link
We were adding the whole GTM JSON to every <a> tag in order to give it a specific category. This is very inefficient as we were duplicating the JSON multiple times. Instead, we can just tag the link with the category we want, as well as any template specific data (indexes). We can then construct the GTM JSON in JavaScript which removes any duplication.