Track clicks on links with child elements
The previous implementation of tracking clicks on links would only check
the tag of the target of the event click. This means that if the link is
a logo containing a span or an image, the click would not be tracked as
it would fail this check. This has been changed to check for the closest
element to be an `a` tag and so it will now track clicks on links with
nested elements.

This issue was first noticed with the GOV.UK logo in the navbar. It had
not been tracking clicks for several months. Changing the check resulted
in the clicks being tracked.