Simplify the way ga4 tracking is added to accordions
Previously we were passing a GTM object to each accordions render statement inside of a frontend app. Instead, it's simpler to just pass a boolean in the frontend apps, and then we handle the GTM object inside of govuk_publishing_components. This gives us the benefit of only having to maintain our GTM code in one repo.

The old show_all_attributes took a JSON which contained JSONS, and appended each one as a data-attribute to the 'Show all link'. This broke the refactored accordion tracking, as we now only pass through one JSON, not a JSON which contains JSONs. The old show_all_attributes is no longer needed for GA4, but we can't modify the code for it as it is now being used for UA. Therefore I've separated the new ga4 functionality out, which as a side effect gives us a more accurate data attribute name, 'data-ga4-show-all-attributes'.