Delete unused classes/CSS from sitewide navbar CSS
Remove the CSS for the previous implementation of the navbar which is
now no longer needed. As part of refactoring of the template we now use
one button for all viewports instead of different buttons at different
viewports. Remove also the CSS for the styling the buttons differently
at mobile viewports.