Change CSS of layout_super_navigation_header
Make changes to the CSS of layout_super_navigation_header to take into
account the new changes to structure and classes used in the template.

These changes to the CSS are; to make the menu that appears at desktop
viewports also appear at mobile viewports; a pseudoelement seperator
line to mark a delineation between the two lists of links; ensuring the
focus styles still look the same as they did previous; and making sure
the link fallback when JS is not running still functions as expected.