Change YML for Navbar
As part of the redesign of the navbar from two dropdown buttons to one
dropdown button, the content of the navbar has changed in strucutre. In
this commit we replace the list of buttons with a single button and add
support for specifying the size of column to use in the single dropdown
menu.