Restructure layout_super_navigiation_header HTML
In the redesign of the site-wide navbar, there will only be one
drop down menu. This means a loop to multiple drop down buttons is
no longer required. In this commit, I have moved the HTML that will
be still used outside of the loop. This will make subsequent changes
easier to discern.