Delete layout_super_navigation_header HTML
Remove HTML that is no longer being used in the sitewide navbar.

Instead of using a different set of buttons at mobile and desktop
viewport, only one button will be used at all viewports. This means
the "mobile only" button can be removed (as well as the link).

In the new redesign of the dropdown menu in the sitewide navbar, there
are no longer any footer links. These have been removed from the .yml
already and this commit removes the HTML.