Change HTML layout_super_navigation_header
This commit changes the HTML to work with the new structure in the .yml.
We are going from two buttons in the navbar, to one and so variable
names have been updated in accordance with this.

The new dropdown menu in the sidewide navbar, consists of columns. This
commit updates the template to use columns for each different group of
links and use the size of column specified in the .yml.

We also change data-attributes so that JS will now work for a single
button in the navbar. These changes mean that there will be correct
margin between the navbar and the page content. It will also mean that
the navbar will remain open between different viewports.