Use the base path for World Location related links
We are currently creating the URL of World Location related links by
using the World Location's name.

However there are multiple examples where the name does not match the
slug (e.g. an acronym in the slug, but spelt out in full in the name).

This was attempted to be solved by hardcoding a list of special cases,
but that list has now become out of date.

Since the content item already includes the base path (which will always
be up-to-date), we should use that and remove the list of exceptions.