Remove inline styling from feedback component
To improve users security we intend to disallow the use of unsafe inline
styles by the content security policy [1].

This means that we need to remove the use of inline style tags like
those below.

From my research these attributes were added as an idea that they may help
reduce spam [2], however this did not resolve the spam problem and
further attempts were made [3]. However this code was left in.

The process to update the CSP will take a while, we will first remove
global offences like the below, then later remove unsafe-inline from the
CSP with report-only on so we can discover if we have other offences and
finally turn on the CSP. This should give us amble time to discover if
this was impactful in the battle against spam bots - my expectation is
that it's impact was negligible.

[1]: https://content-security-policy.com/unsafe-inline/
[2]: https://github.com/alphagov/govuk_publishing_components/pull/2568
[3]: https://github.com/alphagov/govuk_publishing_components/pull/2574