Pure CSS approach to numbering govspeak steps
This changes the approach for styling the custom list item markers for a
Govspeak step list. Previously this was generated with a custom SVG
image generated 30 times. Instead this uses a before pseudo-element with
a content attribute of the number.

The motivation for this change was the upcoming changes to the GOV.UK
Content Security Policy that will disallow the use of data:image due to
the increased security risk these can have [1].

This however has additional benefits. It reduces the size of CSS by
approximately 13 KB and it resolves an ARIA warning in the component
gem.

There were a few considerations made in putting this together:

1. I used the before pseudo-element rather than marker for IE support
2. I used the :before syntax rather than ::before as this seems to be
   the prevalent style in this codebase (presumably an IE legacy
   concern)
3. I couldn't seem to exactly match the weight of the font in the
   previous SVG, it seems it is somewhere between normal and bold
4. I had a fight with IE to avoid some extra padding in the list items,
   which led to the extra commented rules

I've tested it against the following browsers:

- Chrome
- Edge
- Firefox
- Safari
- Opera
- IE 11
- IE 10
- IE 9

[1]: https://github.com/alphagov/govuk_app_config/blob/9b01e06d889be266501714e0dc0fc70a86e7e240/lib/govuk_app_config/govuk_content_security_policy.rb#L33-L37