Replace data:image based test for a CSP friendly one
We are intending to remove both the ability to set inline styles and the
ability to use data images in the GOV.UK Content Security Policy [1]. This
is because both of them offer potential security risks.

However a problem was that this test component relied on both of these.
After much digging I found a different issue that could trigger an
incomplete accessibility warning and I've converted the test to use
that.

[1]: https://github.com/alphagov/govuk_app_config/blob/9b01e06d889be266501714e0dc0fc70a86e7e240/lib/govuk_app_config/govuk_content_security_policy.rb#L32-L65