WIP use specified schemas to resolve flaky tests
The random choice of schema can cause these tests to fail as unexpected
code paths are executed. If the schema chosen is of type
"html_publication" then code will execute that causes an unmocked HTTP
request to occur which fails the test [1].

We can fix this by making the schema choice deterministic.

Note, the test we know to be flaky is the related links one [2], however
I've amended other ones as I have assumed (but not proven) that they're
also flaky for the same issue.

[1]: https://github.com/alphagov/govuk_publishing_components/blob/02d23a11933299d8d56ec2510a012b192b96326d/lib/govuk_publishing_components/presenters/breadcrumb_selector.rb#L23
[2]: https://github.com/alphagov/govuk_publishing_components/blob/main/spec/features/contextual_navigation_spec.rb#L38-L43