Support Rails Content Security Policy nonce on inline JavaScript
This changes the approach to adding inline script to components, the
motivation for this change is to work towards removing the unsafe-inline
attribute from the GOV.UK Content Security Policy [1].

In order to use inline script when unsafe-inline is removed we need
script tags to be annotated with a nonce [2] attribute that matches one
sent in the HTTP header. This helps prevent XSS scripting as injecting
script cannot guess what the nonce would be.

If an app has a Rails nonce configured [3] this will add the nonce
attribute to the script tags. If an app does not have a nonce generator
configured then the script tag will not be given a nonce attribute.

In order to implement this fully to GOV.UK there will also be changes
needed in Slimmer to inject nonces into layouts.

Note a consequence of using the Rails helper for script tags is that
they are wrapped in CDATA escaping, while a bit legacy (it's from the
XHTML days) this has no functional impact.

Before:

```
<script>
  document.body.className = ((document.body.className) ? document.body.className + ' js-enabled' : 'js-enabled');
</script>
```

After (with nonce generator configured):

```
<script nonce="ntkA9+c5yZHnUMvxbOPsTw==">
//<![CDATA[
      document.body.className = ((document.body.className) ? document.body.className + ' js-enabled' : 'js-enabled');

//]]>
</script>
```

[1]: https://github.com/alphagov/govuk_app_config/blob/31928b2079d82f6cbe7c296d4d6e54b4efc07b84/lib/govuk_app_config/govuk_content_security_policy.rb#L55-L59
[2]: https://content-security-policy.com/nonce/
[3]: https://edgeguides.rubyonrails.org/security.html#adding-a-nonce