Use upcoming CSP rules when using component gem
This depends on: https://github.com/alphagov/govuk_publishing_components/pull/3173

This customises the content security policy (CSP) configuration for the
dummy app of the component gem to use the CSP that GOV.UK is working
towards applying.

It is expected to be removed in the near future once the GOV.UK global
CSP applies the same rules, and won't have negative effects if the
GOV.UK CSP is changed without this changing.

This has been done so that developers working with the component gem are
working with the rules we are aiming for to avoid any regressions and learn
of any problems.