Add Youtube Embed to Image Card component
As part of the redesign for the Number 10 organisation page we want to
be able to embed Youtube videos in the image card component.

This commit adds a youtube_video_id parameter to the ImageCardHelper as
well as several functions for generating a youtube link  using a thumbnail
of the given youtube video. This link will replaced by the embed if the
JS is able to run.

Now we have added a Youtube embed to the image card, we need to add some
tests to the image_card spec. This makes sure that we are adding the
data attribute required for the javascript to work and also that the
fallback link is correctly generated from the helper.

We don't test the embed itself because we are simply running the
YoutubeLinkEnhancement module without modifications. This already has
tests written for it.

In this commit, we have also have added the functionality
for turning a link to Youtube into an embed in the image card to
image_card.js. This function uses the existing YoutubeLinkEnhancement
module from Govspeak. There was only one minor modification needed to
be made to YoutubeLinkEnhancement for this to work, which was to allow
specifying a class to be appended to the Youtube embed.

Have also added the data-attribute required for the ImageCard js
module to be initialised in the image card partial if youtube_video_id
has been passed to it.