Style fallback Youtube Embed in Image Card
When a user has JS/cookies disabled, the Youtube embed in image card
will not be generated. Instead the user will see the link to the Youtube
video. This comprises of a thumbnail of the video as well as a text
description of the video. To make it more clear that the link is going
to Youtube, we have overlaid on top of the thumbnail a Youtube icon. To
make it clear that the user has focused the link to the Youtube video,
focus style has been added to the text of the link when the link is
focused.