Initialise GTM in page head instead of after DOMContentLoaded
This change moves all the functionality for loading GTM into load-analytics.js. It will call the code to load GA4 as soon as load-analytics.js is imported. Firstly, it will define the GA4 vars. Then, it will add the GTM script to the head of the page. Then, it will only initialise the GA4 modules once the DOM has loaded. This better follows the best practice of loading GA4 as early as possible. It also fixes a bug in the eCommerce tracking where it was relying on window.dataLayer existing in order to initialise itself. This bug was introduced as we were initialising GA4 after the eCommerce tracking, whereas now it should run first, as adding the GTM script doesn't need to wait for the DOM to load.

Move detecting the GA4 environment in to the main loadGa4 function

Fix load-analytics.spec.js tests and combine the init-ga4.spec.js tests into here now that init-ga4.js no longer exists

Stop UA loading during testing to fix a failing test

Delete ga4-core.js load function and move tests to load-analytics.js

Fix failing tests when GA4 sometimes initialises inside of tests

Changes in response to PR