Add classes to align text in Govspeak tables
This change is made to support the proposed change in:
https://github.com/alphagov/govspeak/pull/268, which is made to allow
for GOV.UK to use a Content Security Policy that can help prevent
Cross-site scripting (XSS) attacks. This change will need to be merged
and released first.

This adds three classes that can be used within `.gem-c-govspeak table`
to set the text alignment of cells. These are required because the
Govspeak output for text aligned cells is going to change from an inline
style (`<td style="text-align: right;">item</td>`) to a class (`<td
class="cell-text-right">item</td>`)

Class naming wise, I didn't use a BEM approach as there did not to be
precedent for it in the classes within govspeak. I then took some
inspiration from tailwind conventions to apply text-<direction> to a
prefix of `cell-`.