Remove margin added by JS in navbar
This is part of the work to make the navbar relatively positioned
instead of absolutely positioned.

As we are going to switch the navbar to being relatively positioned
instead of absolutely positioned, we don't need to add a bottom margin
to the navbar to push the content of the page down when the navbar is
open. The function for this has been removed from js of the layout super
navigation header component and the listener which runs the function on
resize has also been removed.