Allow a tracked link url to be overwritten by its ga4_link JSON
There is a need to track a button element as a link, which doesn't have a href that can be grabbed. Therefore we need to be able to set the URL to track in the ga4_link JSON itself. Therefore, this commit won't search for a href attribute on a link when the URL has been explicitly set on the ga4_link JSON of the thing we click.