Update navbar js spec
Use the new HTML for the updated layout super navigation header in the
JS tests. Remove the screen size specific JS tests because as of the
previous changes (#3058) to the navbar, the navbar buttons are the same
on all viewports. This should have been changed before but as the test
worked, I did not notice.