Force English text on our GA4 related navigation section tracking
These headings get translated, however for the purpose of analysis we always want it to be English, so that everything is tracked under the same label even if translated when presented to the user. This also contains a refactor of construct_section_heading so that it can work with other attributes.

Co-authored-by: Graham Lewis <44037625+gclssvglx@users.noreply.github.com>