Always require AssetHelpers so gem can operate without Rails
This fixes a problem I encountered when running Govspeak with the
current version of the component gem:

```
Error:
GovspeakAttachmentTest#test_renders_an_attachment_component_for_a_found_attachment:
ActionView::Template::Error: undefined method `add_gem_component_stylesheet' for #<ActionView::Base:0x00000000002da0>
```

This error occurs because the views in this gem require the methods from
GovukPublishingComponents::AppHelpers::AssetHelper and if they are only
added if Rails exists then this gem will error. Govspeak is a user of
this gem that does not have rails.

As this app is tested in a Rails environment it's very hard to write a
test that exposes this issue and can cover a regression therefore I've
not added one.