Add data attributes to attachment links before ga4-link-tracker is initialised
Initially we were going to set the GA4 data attributes directly on the attachment link HTML. However, the HTML lives in Whitehall, meaning that in order to update the HTML we would need to republish 45k+ documents. This could become an issue if we constantly need to make changes to the data attributes. Therefore, we are going to set 'data-ga4-attachment-links' on the HTML of the attachment links once. Then we can handle the actual data attributes in JS. This will stop us having to republish 45k+ documents if the data attributes ever need changing/updating.