Ensure Cookie Banner Parity with Design System
The Cookie Banner in govuk_publishing_components has different HTML to
the Cookie Banner in the Design System. This is despite the fact that
the they both resemble each other and have no differences in
functionality. Ideally they should be using the same mark-up and
classes as the Design System implementation is backed up by
user-research and accessibility testing. It will also make it easier
to keep up with future updates to the Cookie Banner.