Move step by step link tracking to use data-ga4-track-links-only
Previously we were only adding the link tracking to links that were inside a list inside a step. However there were also links inside of content. We can't access that content via the code so those weren't being tracked. To fix this, we use the data-ga4-track-links-only technique on the parent element of each step instead. This uses JS so we can track the links in the content as well as in the list.