Cookie Banner Spec Change
As we have changed the partial of the Cookie Banner component to make it
in line with the Design System implementation, we need to update the
tests to make sure they select the correct elements.

Add a new test to make sure that the refocusing works when the user
clicks Accept/Refuse cookies.