Add GA4 tracking object for adding and removing filters
In order to do this, we need to setIndexes on each filter section first. I added a data-ga4-filter-container attribute to the filter container in finder-frontend. I then set a data-ga4-section attribute on each filter which contains the name of the filter. We also use this data-ga4-section to loop through each section and set an index on it. Then we can grab those values to help us create the GA4 object.