Add global-bar class to public_layout
A standard fixture of the public_layout is the blue bar underneath the
navbar. It is fairly standard on pages that do not have a heading section
with a background colour. If a page does have a background colour
however it results in a gap between the blue bar and the header. As a
result there is the option to remove the blue bar from a template in the
event that an application has a full-width heading section with a
background colour (such as the Browse pages on collections). In these
applications, it is then possible to implement the blue bar in the
heading section so that this gap does not occur.

The global bar can be deployed on all pages on GOV.UK. This component
has a full-width blue bar which conflicts with the blue bar
implementation in specific applications.

The simplest solution I could determine was for there to be some way of
an application 'knowing' that a global bar had been deployed. Then in
the application the blue bar could be hidden. So I have added a class
to the body of public_layout so that it is possible to write a rule
which hides an application's blue bar when the global bar is deployed.