Add tracking: contents link list
I didn't use data-ga4-set-indexes as the page you're on is part of the list, but it isn't a link. So it needs to be included in the indexes, but it's not a HTML link. data-ga4-set-indexes only sets indexes by presence of HTML links. Therefore the index_total and indexes would always be off. Therefore we manually set the indexes in the template so that the item that isn't a link in the list is included.