Use actions/checkout for checking out repo rather than git directly
This change has been done for consistency with other places where we
checkout repos. I wasn't able to clone it down a directory as previous
so I had to set an env var - this is consistent with other repos [1].

I'm not too sure if using this approach gives us any advantages but it
seems less surprising.

[1]: https://github.com/alphagov/govuk-developer-docs/blob/cad1554acdff89c3a5efea5621211793b16513c6/.github/workflows/ci.yml#LL37C11-L37C73