Add support for HTML and external attachment types
This commit adds support for HTML and external attachments to the
Attachment component.

So far, the Attachment component has only been able to represent file
attachments – i.e. attachments which have a 'content type' and a
thumbnail that changes depending on the type of file it is.

This change acts as an enabler for us to switch off custom rendering of
embedded attachments in the Whitehall Publisher app. It also ports
across the HTML attachment thumbnail SVG which is currently held in
Whitehall. [1]

I've also taken the opportunity to tighten up the tests to specify which
thumbnail should appear for the various different types of attachments.
This just supplements existing tests rather than explicitly testing them
independently.

I've also fixed a minor styling bug which was present, where custom
thumbnail images didn't receive a grey border because the `<img>`
element was missing the required CSS class name.

[1]: https://github.com/alphagov/whitehall/blob/340d9e54661537f6efe63e91b952eb8289d74920/app/assets/images/attachment-icons/html.svg