Add 'View online' preview link to Attachment
This adds a new `preview_url` parameter that can be passed to the
Attachment component to show a "View online" preview link.

This change acts as an enabler for us to switch off custom rendering of
embedded attachments in the Whitehall Publisher app.

Whitehall uses the 'View online' link to allow users to preview the
contents of CSV files online, without having to download the file to
their device and open it in a compatible application.

Unforuntately we don't have translations for the "View online" label, so
I've left it blank in all non-English locales. We may choose to add
these at a later date.