Improve single page notification button helper
Prior to this commit:
If skip_account is present, the form action for the single page notification button is the
non-govuk-acount signup endpoint and the params necessary for that route are added to the request.

If skip_account is nil, default to the govuk-account form action.

New changes:
Instead of checking if skip_account is present, check if skip_account == "true"

Why?

When this component is called by the frontend, we can pass anything into the the skip_account
field, so we need to make the code here is stricter. For example passing "false"
to skip_account in Frontend, we would expect the component to implement the default non-govuk-account
sign up path for the button. But in fact that would not have been the case, because "false".present?
returns true.